refactor(pep723): move parsing to `utils`
